requirements.txt
torch
torch-geometric
pyyaml
scikit-learn
fastapi
uvicorn
numpy
ogb

